Moved search bar into core components folder.
